// api/_xata.ts
// Generated by Xata Codegen 0.29.5. Please do not edit.
import { buildClient } from "@xata.io/client";
import type {
  BaseClientOptions,
  SchemaInference,
  XataRecord,
} from "@xata.io/client";

const tables = [
  {
    name: "IRCustIDMapping",
    columns: [
      { name: "userID", type: "string", unique: true },
      { name: "irCustID", type: "string" },
      { name: "verifyCode", type: "int" },
      { name: "isVerified", type: "bool", defaultValue: "false" },
      { name: "msgSent", type: "bool", notNull: true, defaultValue: "false" },
    ],
  },
] as const;

export type SchemaTables = typeof tables;
export type InferredTypes = SchemaInference<SchemaTables>;

export type IRCustIDMapping = InferredTypes["IRCustIDMapping"];
export type IRCustIDMappingRecord = IRCustIDMapping & XataRecord;

export type DatabaseSchema = {
  IRCustIDMapping: IRCustIDMappingRecord;
};

const DatabaseClient = buildClient();

const defaultOptions = {
  databaseURL:
    "https://arturo-mayorga-s-workspace-3e5jp6.us-east-1.xata.sh/db/bfr-leapdb",
};

export class XataClient extends DatabaseClient<DatabaseSchema> {
  constructor(options?: BaseClientOptions) {
    super({ ...defaultOptions, ...options }, tables);
  }
}

let instance: XataClient | undefined = undefined;

export const getXataClient = () => {
  if (instance) return instance;

  instance = new XataClient();
  return instance;
};